+++ mkdocs/docs/about/release-notes.md
@@ -1,12 +1,21 @@
 # Release Notes
 
+## Version 0.0.3
+
+* `Anima`: The `Looker.py` moved into Package and create
+* `Anima`: `Oracle.py` class responsible by predictions models
+* `Assistant`: Intro `Scissor.py` to cut images. `API.py` moved into this Package
+* `Assistant`: `Reader.py` improved to make a place – without city information – a fictional city
+* `Hive`: Agents now drive between two dots
+* `System`: improved the `Graphium.py` to scissor information
+
 ## Version 0.0.2
 
-* `Assistant`: Intro `GeoSpatial.py` to cal distance and intermediate points of sistem
+* `Assistant`: Intro `GeoSpatial.py` to cal distance and intermediate points of system
 * `Assistant`: Intro `Looker.py` search the pictures on google street view
 * `Hive`: Improved the `Agent.py` and `API.py`
 * `System`: Now the `Graphium.py` has the config to download images from Google Street View
-* `System`: Graphium receive the picture folder to not request two time the same pano image. Google restringe 25.000 requests by day
+* `System`: Graphium receive the picture folder to not request two time the same pano image. Google restricts 25.000 requests by day
 
 ## Version 0.0.1